fix(runtime): resolve method + body symbols, keep error.rs under the LOC cap (#9486)
Register the display name against the `perry_fn_*` body symbol and
against `perry_method_*`, since a direct call targets the body rather
than the closure-value wrapper. Move the capture/format helpers into
error_stack_frames.rs (error.rs was 2043 lines, cap is 2000).
